+++ packages/store/plugins/src/index.ts
@@ -1,3 +1,9 @@
 export { InitState, UpdateState } from './actions';
-export { NGXS_PLUGINS, NgxsPlugin, NgxsPluginFn, NgxsNextPluginFn } from './symbols';
+export {
+  NGXS_PLUGINS,
+  NgxsPlugin,
+  NgxsPluginFn,
+  NgxsNextPluginFn,
+  ɵisPluginClass
+} from './symbols';
 export { getActionTypeFromInstance, actionMatcher, setValue, getValue } from './utils';