+++ packages/store/src/standalone-features/plugin.ts
@@ -5,7 +5,7 @@ import {
   inject,
   makeEnvironmentProviders
 } from '@angular/core';
-import { NGXS_PLUGINS, NgxsPlugin } from '@ngxs/store/plugins';
+import { NGXS_PLUGINS, NgxsPlugin, NgxsPluginFn, ɵisPluginClass } from '@ngxs/store/plugins';
 
 import { PluginManager } from '../plugin-manager';
 
@@ -23,12 +23,18 @@ import { PluginManager } from '../plugin-manager';
  * });
  * ```
  */
-export function withNgxsPlugin(plugin: Type<NgxsPlugin>): EnvironmentProviders {
+export function withNgxsPlugin(plugin: Type<NgxsPlugin> | NgxsPluginFn): EnvironmentProviders {
   return makeEnvironmentProviders([
-    { provide: NGXS_PLUGINS, useClass: plugin, multi: true },
+    ɵisPluginClass(plugin)
+      ? { provide: NGXS_PLUGINS, useClass: plugin, multi: true }
+      : { provide: NGXS_PLUGINS, useValue: plugin, multi: true },
     // We should inject the `PluginManager` to retrieve `NGXS_PLUGINS` and
     // register those plugins. The plugin can be added from inside the child
     // route, so the plugin manager should be re-injected.
-    { provide: ENVIRONMENT_INITIALIZER, useValue: () => inject(PluginManager), multi: true }
+    {
+      provide: ENVIRONMENT_INITIALIZER,
+      useValue: () => inject(PluginManager),
+      multi: true
+    }
   ]);
 }